optimizations, restructuring and mouse option to bonus rule added
