+++ main.c
@@ -6,7 +6,7 @@
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/01 19:17:12 by skunert           #+#    #+#             */
-/*   Updated: 2023/11/04 13:11:21 by njantsch         ###   ########.fr       */
+/*   Updated: 2023/11/04 20:38:44 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
