+++ maps/test.cub
@@ -1,4 +1,4 @@
-			C	  255 	 	 , 	 	0 	 	 ,	 	 0
+			C	  200 	 	 , 	 	130 	 	 ,	 	 11
 
 EA          ./textures/wooden_wall_no_window.png
 
@@ -15,32 +15,70 @@ NO              ./textures/wooden_wall.png
            WE        ./textures/wooden_wall_no_window.png
 
 
-		F       0 	 	 , 	 	0 	 	 	 	 	,    		255
+		F       50 	 	 , 	 	160 	 	 	 	 	,    		122
 
+                                                        111111
+                                                        11000111111111
+                11111111111111111                       11000000000001
+                100000000000000011111111111111111111111111000000000001
+                100000000000000000000000000000000000000000000000000001
+                100000000000000011111111111111111001111111111111011111
+                10000000000000001               1001           101
+                10000000000000001               1001           101111111111111111
+                10000000N00000001               1001           100000000000000001
+                10000000000000001           111110011111       111111111111111101
+                10000000000000001           100000000001                      101
+                10000000000000001           100000000001                      101
+                10000000000000001111111111111000000000011111111111111111111111101
+                10000000000000000000000000000000000000000000000000000000000000001
+                11111011111111111111111111111000000000011111111111111111111111101
+                    101                     100000000001                      101
+                11111011111111111           100000000001                      101
+                10000000000000001           111110011111                      101
+                10000000000000001               1001                          101
+                10000000000000001111111111111111100111111111111111111111111111101
+                10000000000000000000000000000000000000000000000000000000000000001
+                10000000000000001111111111111111111111111111111111111111111111111
+                10000000000000001
+                10000000000000001
+                10000000000000001
+                10000000000000001
+                10000000000000001111111111111
+                10000000000000000000000000001
+                11111111111111111111111111101
+                                         11011
+                                        11000111111111
+11111111111111111                       11000000000001
+100000000000000011111111111111111111111111000000000001
+100000000000000000000000000000000000000000000000000001
+100000000000000011111111111111111001111111111111011111
+10000000000000001               1001           101
+10000000000000001               1001           101111111111111111
+10000000000000001               1001           100000000000000001
+10000000000000001           111110011111       111111111111111101
+10000000000000001           100000000001                      101
+10000000000000001           100000000001                      101
+10000000000000001111111111111000000000011111111111111111111111101111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111
+100000000000000000000000000000000L4000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001
+11111011111111111111111111111000000000011111111111111111111111101111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111
+    101                     100000000001                      101
+11111011111111111           100000000001                      101
+10000000000000001           111110011111            1         101
+10000000000000001               1001                          101
+10000000000000001111111111111111100111111111111111111111111111101
+10000000000000000000000000000000000000000000000000000000000000001
+10000000000000001111111111111111111111111111111111111111111111111
+10000000000000001
+10000000000000001
+10000000000000001
+10000000000000001
+100000Z0000000001
+10000000000000001
 11111111111111111
-10000000000000001
-10000000000000001
-10000000000000001
-10000000000000001
-10000000000000001
-10000000N00000001
-10000000000000001
-10000000000000001
-10000000000000001
-10000000000030001
-10000000000000001
-11111011111111111
-    101
-11111011111111111
-10000000000000001
-10000000002000001
-10000000000000001
-10200000000200001
-10000000000000001
-10000000000200001
-10200000000000001
-10000000400200001
-10000000000000001
-1000000Z000200001
-1020000000L000001
-11111111111111111
+111111
+1111
+111
+11
+1
+1
+1
