+++ srcs/graphics/set_sky_floor_color.c
