+++ srcs/minimap/draw_line.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   draw_line.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: skunert <skunert@student.42heilbronn.de    +#+  +:+       +#+        */
+/*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/09 17:22:42 by njantsch          #+#    #+#             */
-/*   Updated: 2023/10/26 11:29:52 by skunert          ###   ########.fr       */
+/*   Updated: 2023/11/04 20:01:50 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ void	ft_draw_line(t_game *game, int i)
 			break ;
 		i = -1;
 		while (++i < 3)
-			mlx_put_pixel(game->minimap, game->dl->begin_x + i,
+			mlx_put_pixel(game->minimap_rc, game->dl->begin_x + i,
 				game->dl->begin_y, 0xF0F00FFF);
 		game->dl->two_times_error = 2 * game->dl->error;
 		if (game->dl->two_times_error > -game->dl->delta_y)
