+++ srcs/raycaster/caster.c
@@ -6,7 +6,7 @@
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/06 17:08:51 by njantsch          #+#    #+#             */
-/*   Updated: 2023/11/04 13:46:16 by njantsch         ###   ########.fr       */
+/*   Updated: 2023/11/04 20:08:12 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,12 +19,17 @@ void	replace_img(t_game *g)
 		mlx_delete_image(g->mlx, g->minimap);
 	if (g->line)
 		mlx_delete_image(g->mlx, g->line);
+	if (g->minimap_rc)
+		mlx_delete_image(g->mlx, g->minimap_rc);
 	g->line = mlx_new_image(g->mlx, g->dis_w + 20, g->dis_h + 20);
 	g->minimap = mlx_new_image(g->mlx, g->dis_w, g->dis_h);
+	g->minimap_rc = mlx_new_image(g->mlx, g->dis_w, g->dis_h);
 	mlx_image_to_window(g->mlx, g->line, 0, 0);
 	mlx_image_to_window(g->mlx, g->minimap, 20, 20);
+	mlx_image_to_window(g->mlx, g->minimap_rc, 20, 20);
 	g->line->instances[0].z = 0;
 	g->minimap->instances[0].z = 1;
+	g->minimap_rc->instances[0].z = 2;
 }
 
 // sets limits for rotation
@@ -73,7 +78,6 @@ void	raycaster(t_game *g)
 	g->ray->ray_a = g->caster->pa - DGREE * 30;
 	set_limit(g);
 	replace_img(g);
-	ft_fill_minimap(g);
 	if (SKY == 1)
 		draw_sky(g);
 	while (++g->ray->rays < g->ray->n_of_rays)
@@ -87,5 +91,5 @@ void	raycaster(t_game *g)
 	ft_set_values_sprites(g);
 	ft_set_values_luffy(g);
 	ft_set_values_zorro(g);
-	ft_fill_minimap_walls(g);
+	ft_fill_minimap(g);
 }
