+++ srcs/raycaster/horizontal.c
@@ -6,7 +6,7 @@
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/10 21:26:51 by njantsch          #+#    #+#             */
-/*   Updated: 2023/11/03 21:02:12 by njantsch         ###   ########.fr       */
+/*   Updated: 2023/11/04 18:14:39 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
