+++ srcs/raycaster/vertical.c
@@ -6,7 +6,7 @@
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/10 21:27:23 by njantsch          #+#    #+#             */
-/*   Updated: 2023/11/03 21:02:26 by njantsch         ###   ########.fr       */
+/*   Updated: 2023/11/04 18:14:27 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
